Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (7 comments)
> 
> Hi Damian,
> 
> thanks for the patch. Here's a first pass of comments.

Hi Jan-Simon,
Thanks for the comments.

I will package the tests into a -ptest.
Do you have any ideas about which layer the fff recipe should go in?
Right now, the fff.wrap will only be used if it can't be found locally (i.e. the recipe-sysroot), so
if a fff recipe is available it will automatically use that version.

Patch-set: 1