Don't close invalid or uninitialied fds
Avoid calling close() on file descriptors that have not been intialized,
(or initialized to 0) and descriptors that are known to be invalid.

Closing file descriptors due to handle failing initialization is handled
in the intialization code. In the lease-manager, file descriptors can be
invalid during normal operation, so validity is tested at shutdown.

Bug-AGL: SPEC-3862

Change-Id: I798273195cba297c14c6b97d50c7614164fda7df
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>