test/drm-lease: Create helper functions to reduce boilerplate
Most of the test cases use very similar setup code to configure
the test DRM device and leases.  Refactor this out into separate
functions to help keep the focus on the relevant parts of each
test.

The helpers set up a basic DRM device with one 1:1 mapping of
encoders, CRTCs, and connectors, with support for planes.

Tests that don't adhere to this pattern are likely testing for a
specific configuration, so should be open coding their setup.

Bug-AGL: SPEC-3815

Change-Id: I0adbed04246ff6af0060692fa4c1e5897b903237
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>