Multi connector lease support
Add support for having multiple connectors / CTRCs in a single lease.
The lease manager receives a list of lease configurations that
specify the name and list of connectors to include in each lease.

All DRM planes that are not shared between muliple CRTS are added
for each connector contained in the lease.

If the lease manager is initialized without a list of lease
configurations it will fall back to the previous 1 connector
per lease configuration.

Bug-AGL: SPEC-3815

Change-Id: I6699e7c8f426bab9fadbc64fe4ed242a834376bf
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>