Add plane setting to connector configuration
Provide a setting to override the default DRM plane
alloction for each connector in a lease.

DRM planes that are compatible with multiple CRTCs/
connectors can now be assigned to a lease by
explicity setting the connector that they belong to
in the configuration file.

This can also be used to restrict the planes that
are available to a given connector.

Bug-AGL: SPEC-3815

Change-Id: If4168fc3081213c4ed1df09755cadf042c68d9a3
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>