Add 'optional' property to connector configuration
All connectors will default to mandatory. i.e. if
any specified connector is in the lease configuration
is not available, that lease will not be created.
Setting a connector as 'optional' lets the lease
creation succeed even if the connector is not
physically present on the system.

Failing to create a lease does not affect the creation
of other leases.  The drm-lease-manager daemon will
run as long as one lease is successfully created.

Bug-AGL: SPEC-3815

Change-Id: I5edf8a97a2a3589e8eb5368c0a5b13adb4cb5c9b
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>