Force add a primary plane to lease
A primary plane of crtc is possible to use in every crtcs in
the configuration of rockchip drm driver.

Planes:
id crtc fb  CRTC x,y x,y gamma size possible crtcs
54 82   451 0,0	     0,0 0          0x0000000f

In this case, existing drm lease manager judges to shared plane
that is not primary plane.
This patch add to fixup loginc in this case.

Bug-AGL :

Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>